+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # coding: utf-8
 $:.unshift File.expand_path("../lib", __FILE__)
 require 'rubygems'
+require 'pry'
 
 require "codeclimate-test-reporter"
 CodeClimate::TestReporter.start